feat(jev): modernize JEV form with inline grid editing and validation
- Replace modal account entry forms with inline DataGridView editing
- Refactor ucJEV to use dynamic combo columns (FPP, Account, Subsidiary)
- Implement real-time debit/credit balancing with visual indicator
- Add robust cell-level validation and error reporting
- Update data model mapping to use visible column values (fpp, gen_ldgr_acc, etc.)
- Remove deprecated hidden columns (FPPId, GeneralLedgerId, etc.)
- Adjust UI layout and styling for better usability
- Standardize repository method signatures to use int instead of byte/ushort
- Add SubmitJev entry point for unified insert/update logic

+++ ACC.Data/SubsidiaryLedgerAccountsRepository.cs
@@ -116,12 +116,12 @@ namespace ACC.Data
             return mySqlGenericCommandsLFS.FillBySearch(query, dataTable, parameters);
         }
 
-        public DataTable GetRecordsByFundAndGeneralLedger(byte fundId, ushort generalLedgerId)
+        public DataTable GetRecordsByFundAndGeneralLedger(int fundId, int generalLedgerId)
         {
             object[][] parameters = new object[][]
             {
-                new object[] { "@funds_id", DbType.Byte, fundId},
-                new object[] { "@generalLedgerId", DbType.UInt16, generalLedgerId},
+                new object[] { "@funds_id", DbType.Int32, fundId},
+                new object[] { "@generalLedgerId", DbType.Int32, generalLedgerId},
             };
 
             string query = $"SELECT * FROM {tableName} WHERE funds_id = @funds_id AND general_ledger_accounts_id = @generalLedgerId";
